refactor: re-using MockTx
