+++ yarn-project/p2p/src/client/mocks.ts
@@ -1,18 +1,4 @@
-import { KERNEL_PUBLIC_CALL_STACK_LENGTH, Proof } from '@aztec/circuits.js';
-import { makeKernelPublicInputs, makePublicCallRequest } from '@aztec/circuits.js/factories';
-import { Tx, TxL2Logs, L2BlockSource, L2Block } from '@aztec/types';
-import times from 'lodash.times';
-
-export const MockTx = () => {
-  return Tx.createTx(
-    makeKernelPublicInputs(),
-    new Proof(Buffer.alloc(0)),
-    TxL2Logs.random(8, 3), // 8 priv function invocations creating 3 encrypted logs each
-    TxL2Logs.random(11, 2), // 8 priv + 3 pub function invocations creating 2 unencrypted logs each
-    [],
-    times(KERNEL_PUBLIC_CALL_STACK_LENGTH, makePublicCallRequest),
-  );
-};
+import { L2Block, L2BlockSource } from '@aztec/types';
 
 /**
  * A mocked implementation of L2BlockSource to be used in p2p tests.
