+++ yarn-project/p2p/src/client/p2p_client.test.ts
@@ -1,10 +1,9 @@
 import { expect, jest } from '@jest/globals';
-import { L2BlockSource } from '@aztec/types';
+import { L2BlockSource, MockTx } from '@aztec/types';
 
 import { P2PClient } from './p2p_client.js';
 import { TxPool } from '../tx_pool/index.js';
 import { MockBlockSource } from './mocks.js';
-import { MockTx } from './mocks.js';
 import { P2PService } from '../index.js';
 
 /**
