+++ yarn-project/p2p/src/tx_pool/tx_pool.test.ts
@@ -1,4 +1,4 @@
-import { MockTx } from '../client/mocks.js';
+import { MockTx } from '@aztec/types';
 import { InMemoryTxPool } from './index.js';
 
 describe('In-Memory TX pool', () => {
